dependencies added sass loader scss
