+++ src/App.js
@@ -3,24 +3,27 @@ import logo from './logo.svg';
 import './App.css';
 
 function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          OpenBarber by Gabriel Martinez
-        </a>
-      </header>
-    </div>
-  );
+    return ( <
+        div className = "App" >
+        <
+        header className = "App-header" >
+        <
+        img src = { logo }
+        className = "App-logo"
+        alt = "logo" / >
+        <
+        p >
+        Edit < code > src / App.js < /code> and save to reload. <
+        /p> <
+        a className = "App-link"
+        href = "https://reactjs.org"
+        target = "_blank"
+        rel = "noopener noreferrer" >
+        OpenBarber by Gabriel Martinez <
+        /a> <
+        /header> <
+        /div>
+    );
 }
 
 export default App;
