+++ src/App.scss
@@ -1,0 +1,100 @@
+open-chair-app {
+  --app-bg-color: rgb(210, 0, 0);
+  --padding: 8px;
+  margin: 0%;
+  padding: 0%;
+  padding-left: 0%;
+  padding-right: 0%;
+  body {
+    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Fira Sans",
+      "Droid Sans", "Helvetica Neue", sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+  }
+
+  code {
+    font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace;
+  }
+
+  .App {
+    text-align: center;
+  }
+
+  .App-logo {
+    /* animation: App-logo-spin infinite 10s linear; */
+    height: 50px;
+    width: 50px;
+    pointer-events: none;
+  }
+
+  .App-spinner {
+    animation: App-logo-spin infinite 10s linear;
+    height: 50px;
+    width: 50px;
+    pointer-events: none;
+  }
+
+  .App-header {
+    background-color: #282c34;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: calc(10px + 2vmin);
+    color: white;
+  }
+
+  @keyframes App-logo-spin {
+    from {
+      transform: rotate(0deg);
+    }
+    to {
+      transform: rotate(360deg);
+    }
+  }
+
+  .navbar {
+    background-color: var(--app-bg-color);
+  }
+
+  #map {
+    width: 100px;
+    height: 400px;
+    min-height: 100%;
+    min-width: 100%;
+    display: block;
+    padding: 0%;
+    margin: auto;
+  }
+
+  .container-fluid {
+    padding-left: 0%;
+    padding-right: 0%;
+  }
+
+  .alert {
+    visibility: hidden;
+    padding-left: var(--padding);
+    padding-right: var(--padding);
+  }
+
+  #spinner {
+    visibility: visible;
+    height: 100px;
+    width: 100px;
+  }
+
+  /* .tab-pane{
+  width: 100%;
+  height: 100%;
+  padding: 0%;
+  margin: 0%;
+  position:relative;
+}*/
+
+  .row {
+    margin: 0%;
+    padding: 0%;
+  }
+}
