+++ src/index.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 //import App from './App';
 import * as serviceWorker from './serviceWorker';
-import OpenBarber from './open-barber';
+import OpenChair from './open-chair';
 
-ReactDOM.render(<OpenBarber />, document.getElementById('root'));
+ReactDOM.render(<OpenChair />, document.getElementById('root'));
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
